CryptoVisionWebApp/crypto_vision/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
binance-connector==3.5.1
bitquery-python==0.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
crispy-bootstrap5==2023.10
cycler==0.12.1
dateparser==1.2.0
Django==5.0
django-crispy-forms==2.1
djangogrpcframework==0.2.1
djangorestframework==3.14.0
fonttools==4.47.0
frozenlist==1.4.1
grpc-interceptor==0.15.4
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
mplfinance==0.12.10b0
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
pycryptodome==3.19.1
PyJWT==2.8.0
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
